requirements.txt
streamlit
numpy
pandas
scikit-learn
xgboost
joblib
pillow
speechrecognition
plotly
opencv-python-headless
easyocr
torch
torchvision